--- web/src/stories/Button/Button.ts
@@ -1,34 +1,0 @@
-export interface ButtonProps {
-  category: 'primary' | 'secondary' | 'tertiary' | 'negative'
-
-  small?: boolean
-  /**
-   * Button contents
-   */
-  label: string
-  /**
-   * Optional click handler
-   */
-  onClick?: () => void
-}
-
-/**
- * Primary UI component for user interaction
- */
-export const createButton = ({
-  category = 'primary',
-  small = false,
-  label,
-  onClick
-}: ButtonProps) => {
-  const btn = document.createElement('button')
-  btn.type = 'button'
-  btn.innerText = label
-  if (onClick) {
-    btn.addEventListener('click', onClick)
-  }
-
-  btn.className = [`bttn-${category}`, small ? 'small' : null].join(' ')
-
-  return btn
-}
